move text styles to common
