+++ src/common/ui/Text/index.ts
@@ -1,1 +1,10 @@
-export { TextMessage } from "./Text";
+export {
+  TextMessage,
+  TextTitleCardEvent,
+  TextTitleHeader,
+  TextDescriptionCardEvent,
+  TextTableTitleCardEvent,
+  TextTableRowCardEvent,
+  TextTable,
+  TextPriceTableProducts,
+} from "./Text";
