+++ src/partner/modules/CardOrder/OptionsCardOrder.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { CardIconImg } from "src/common/ui/Card";
 import { ArrowOptionsIconImg } from "./CardOrder";
 import styles from "styled-components";
-import { TextPriceTableProducts } from "../ui/Text";
+import { TextPriceTableProducts } from "src/common/ui/Text";
 
 const FloatDivOption = styles.div`
     position: absolute;
