+++ src/partner/modules/Header/NavHeader.tsx
@@ -6,8 +6,7 @@ import {
   HeaderUserImgContainer,
 } from "src/partner/modules/ui";
 import { CardIconImg } from "src/common/ui/Card";
-import styles from "styled-components";
-import { TextTitleHeader } from "../ui/Text";
+import { TextTitleHeader } from "src/common/ui/Text";
 import { LinkStyled } from "src/common/ui/Link";
 
 export interface IHeaderProps {
