--- src/partner/modules/ui/Text/index.ts
@@ -1,9 +1,0 @@
-export {
-  TextTitleHeader,
-  TextTitleCardEvent,
-  TextTableTitleCardEvent,
-  TextTableRowCardEvent,
-  TextDescriptionCardEvent,
-  TextTable,
-  TextPriceTableProducts,
-} from "./Text";
